Refactor and interfaces

+++ src/StandardsSearchIndexer/Sfa.Eds.Standards.Indexer.AzureWorkerRole/Services/StandardService.cs
@@ -1,24 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Configuration;
-using System.Diagnostics;
-using System.IO;
 using System.Linq;
-using System.Net;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.Azure;
-using Microsoft.WindowsAzure.ServiceRuntime;
-using Sfa.Eds.Standards.Indexer.AzureWorkerRole.Models;
 using Nest;
 using Newtonsoft.Json;
 using Sfa.Eds.Standards.Indexer.AzureWorkerRole.Helpers;
+using Sfa.Eds.Standards.Indexer.AzureWorkerRole.Models;
+using Sfa.Eds.Standards.Indexer.AzureWorkerRole.Settings;
 
 namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
 {
     public class StandardService
     {
+        private static readonly StandardIndexSettings StandardIndexSettings = new StandardIndexSettings();
         private static ElasticClient _client = new ElasticClient();
 
         public static async void CreateScheduledIndex(DateTime scheduledRefreshDateTime)
@@ -27,14 +22,17 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
 
             var newIndexName = GetIndexNameAndDateExtension(indexAlias, scheduledRefreshDateTime);
 
-            var node = new Uri(ConfigurationManager.AppSettings["SearchHost"]);
+            var node = new Uri(StandardIndexSettings.SearchHost);
 
             var connectionSettings = new ConnectionSettings(node, newIndexName);
 
             _client = new ElasticClient(connectionSettings);
 
             var existingPreviousIndex = CreateIndex(newIndexName);
-            if (existingPreviousIndex) return;
+            if (existingPreviousIndex)
+            {
+                return;
+            }
 
             await IndexStandards(newIndexName);
 
@@ -48,7 +46,7 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
 
         private static string GetIndexAlias()
         {
-            return ConfigurationManager.AppSettings["StandardIndexesAlias"];
+            return StandardIndexSettings.StandardIndexesAlias;
         }
 
         private static string GetIndexNameAndDateExtension(string indexAlias, DateTime dateTime)
@@ -60,7 +58,7 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
         {
             var indexExistsResponse = _client.IndexExists(indexName);
 
-            //If it already exists and is empty, let's delete it.
+            // If it already exists and is empty, let's delete it.
             if (indexExistsResponse.Exists)
             {
                 var totalResults = _client.Count<StandardDocument>(c =>
@@ -69,14 +67,14 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
                     return c;
                 });
 
-                //if (totalResults.Count == 0)
-                //{
+                if (totalResults.Count == 0)
+                {
                     _client.DeleteIndex(indexName);
                     indexExistsResponse = _client.IndexExists(indexName);
-                //}
+                }
             }
 
-            //create index
+            // create index
             if (!indexExistsResponse.Exists)
             {
                 CreateDocumentIndex(indexName);
@@ -85,50 +83,50 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
             return indexExistsResponse.Exists;
         }
 
-        static public IIndicesOperationResponse CreateDocumentIndex(string indexName)
+        private static void CreateDocumentIndex(string indexName)
         {
-            return _client.CreateIndex(indexName, c => c.AddMapping<StandardDocument>(m => m.MapFromAttributes()));
+            _client.CreateIndex(indexName, c => c.AddMapping<StandardDocument>(m => m.MapFromAttributes()));
         }
 
-        static public async Task UploadStandardsContentToAzure(List<JsonMetadataObject> standardList)
+        private static async Task UploadStandardsContentToAzure(List<JsonMetadataObject> standardList)
         {
             foreach (var standard in standardList)
             {
-                //await UploadStandardJson(standard);
+                // await UploadStandardJson(standard);
                 await UploadStandardPdf(standard);
             }
         }
 
-        static public async Task UploadStandardJson(JsonMetadataObject standard)
+        private static async Task UploadStandardJson(JsonMetadataObject standard)
         {
-            BlobStorageHelper bsh = new BlobStorageHelper();
+            var bsh = new BlobStorageHelper();
             await bsh.UploadStandardAsync("standardsjson", string.Format(standard.Id.ToString(), ".txt"), JsonConvert.SerializeObject(standard));
         }
 
-        static public async Task UploadStandardPdf(JsonMetadataObject standard)
+        private static async Task UploadStandardPdf(JsonMetadataObject standard)
         {
-            BlobStorageHelper bsh = new BlobStorageHelper();
+            var bsh = new BlobStorageHelper();
             await bsh.UploadPdfFromUrl("standardspdf", string.Format(standard.Id.ToString(), ".txt"), standard.Pdf);
         }
 
-        static public async Task<List<JsonMetadataObject>> GetStandardsFromAzure()
+        private static List<JsonMetadataObject> GetStandardsFromAzure()
         {
-            BlobStorageHelper bsh = new BlobStorageHelper();
+            var bsh = new BlobStorageHelper();
             var standardsList = bsh.ReadStandards("standardsjson");
 
             standardsList = standardsList.OrderBy(s => s.Id).ToList();
-            
+
             return standardsList;
         }
 
         private static async Task IndexStandardPdfs(string indexName, List<JsonMetadataObject> standards)
         {
-            //index the items
+            // index the items
             foreach (var standard in standards)
             {
                 var doc = await CreateDocument(standard);
 
-                _client.Index<StandardDocument>(doc);
+                _client.Index(doc);
 
                 /*
                 _client.Index(doc, i => i
@@ -140,7 +138,7 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
 
         private static async Task IndexStandards(string newIndexName)
         {
-            var standards = await GetStandardsFromAzure();
+            var standards = GetStandardsFromAzure();
 
             await UploadStandardsContentToAzure(standards);
 
@@ -154,9 +152,9 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
             }
         }
 
-        static public async Task<StandardDocument> CreateDocument(JsonMetadataObject standard)
+        private static async Task<StandardDocument> CreateDocument(JsonMetadataObject standard)
         {
-            BlobStorageHelper bsh = new BlobStorageHelper();
+            var bsh = new BlobStorageHelper();
 
             var attachment = new Attachment
             {
@@ -191,23 +189,21 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
             return searchResults.Any();
         }
 
-        private static bool SwapIndexes(DateTime scheduledRefreshDateTime)
+        private static void SwapIndexes(DateTime scheduledRefreshDateTime)
         {
             var indexAlias = GetIndexAlias();
             var newIndexName = GetIndexNameAndDateExtension(indexAlias, scheduledRefreshDateTime);
 
             var existingIndexesOnAlias = _client.GetIndicesPointingToAlias(indexAlias);
-            var aliasRequest = new AliasRequest { Actions = new List<IAliasAction>() };
+            var aliasRequest = new AliasRequest {Actions = new List<IAliasAction>()};
 
             foreach (var existingIndexOnAlias in existingIndexesOnAlias)
             {
-                aliasRequest.Actions.Add(new AliasRemoveAction { Remove = new AliasRemoveOperation { Alias = indexAlias, Index = existingIndexOnAlias } });
+                aliasRequest.Actions.Add(new AliasRemoveAction {Remove = new AliasRemoveOperation {Alias = indexAlias, Index = existingIndexOnAlias}});
             }
 
-            aliasRequest.Actions.Add(new AliasAddAction { Add = new AliasAddOperation { Alias = indexAlias, Index = newIndexName } });
+            aliasRequest.Actions.Add(new AliasAddAction {Add = new AliasAddOperation {Alias = indexAlias, Index = newIndexName}});
             _client.Alias(aliasRequest);
-
-            return true;
         }
     }
 }
